requirements.txt
torch>=1.1.0
torchvision>=0.3.0
torchaudio>=0.3.0
h5py>=2.9.0
scipy>=1.3.0
Pillow>=6.0.0
matplotlib>=3.1.0
tensorboard>=2.0.0
opencv-python>=4.0.0
numpy>=1.16.0
pandas>=0.25.0
tqdm>=4.32.0
argparse